+++ supervisely/nn/benchmark/evaluation/coco/__init__.py
@@ -1,0 +1,1 @@
+from supervisely.nn.benchmark.evaluation.coco.calculate_metrics import calculate_metrics
